test/jaytwo.Common.ParseExtensions.UnitTests/Parsers/BoolParsing/OneZeroBoolParserTests.cs
using jaytwo.Common.ParseExtensions.Parsers.BoolParsing;
using System;
using System.Collections.Generic;
using System.Text;
using Xunit;

namespace jaytwo.Common.ParseExtensions.UnitTests.Parsers.BoolParsing
{
    public class OneZeroBoolParserTests
    {
        [Theory]
        [InlineData("1", true)]
        [InlineData("0", false)]
        void Parse_happy_path(string value, bool expected)
        {
            // arrange
            var parser = new OneZeroBoolParser();

            // act
            var result = parser.Parse(value);

            // assert
            Assert.Equal(expected, result);
        }

        [Theory]
        [InlineData("1", true, true)]
        [InlineData("0", true, false)]
        [InlineData("z", false, false)]
        void TryParse_happy_path(string value, bool expectedSuccess, bool expectedResult)
        {
            // arrange
            var parser = new OneZeroBoolParser();

            // act
            var success = parser.TryParse(value, out bool result);

            // assert
            Assert.Equal(expectedSuccess, success);
            Assert.Equal(expectedResult, result);
        }
    }
}
